add AB for cohorts in dedicated controller
